Move common.writebacker_models to writebacker.writebacker_base
Writebacker is neither models nor common. It is at-par with all other functionality like fetcher/syncer/matcher.
